fix timezone env TZ
For e.g, with $TZ="Asia/Chongqing" on host

"TZ=" + time.tzname[0]

will set TZ=CST, which restult in utils such as 'date' output UTC time instead of
CST time because "CST" isn't a valid timezone.

This fix just forwards the $TZ to givsor environment only if it is set on host.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>